Remove managed identity support from WithClaimsFromClient (confidential-client only)
WithClaimsFromClient was added in 4.84.2 for both managed identity and
confidential client flows. This removes the managed identity portion so the
feature is confidential-client only.

BREAKING CHANGE: The
AcquireTokenForManagedIdentityParameterBuilder.WithClaimsFromClient(string)
overload has been removed from the public API surface (including
PublicAPI.Shipped.txt), and managed identity no longer forwards
client-originated claims to IMDS.

Removed (MSI-specific):
- AcquireTokenForManagedIdentityParameterBuilder.WithClaimsFromClient(string)
- AcquireTokenForManagedIdentityParameters.ClientClaims
- ManagedIdentityAuthRequest client-claims propagation
- AbstractManagedIdentity: RequiresMsiV1ClaimsValidation, XmsAzNwperimid,
  the client-claims forwarding block, and ValidateMsiv1Claims
- ImdsManagedIdentitySource.RequiresMsiV1ClaimsValidation override
- AuthenticationRequestParameters.ClientClaims (dead after MSI removal)
- The MSI line from all six PublicAPI.Shipped.txt files
- MSI client-claims unit tests (ImdsV2Tests region and WithClientClaimsTests.cs)

Kept (confidential client / shared):
- WithClaimsFromClient extension on confidential client builders
- AcquireTokenCommonParameters.ClientClaims and the claims/capabilities merge
- The IMDSv2 mTLS-PoP delegation path (works without client claims)

Confidential-client coverage moved to PublicApiTests/WithClaimsFromClientTests.cs.
