+++ src/client/Microsoft.Identity.Client/ApiConfig/Parameters/AcquireTokenForManagedIdentityParameters.cs
@@ -23,12 +23,6 @@ namespace Microsoft.Identity.Client.ApiConfig.Parameters
 
         public string Claims { get; set; }
 
-        /// <summary>
-        /// Client-originated claims to be sent to the identity endpoint.
-        /// Unlike <see cref="Claims"/> (server-issued), these are cached and keyed on the claims value.
-        /// </summary>
-        public string ClientClaims { get; set; }
-
         public string RevokedTokenHash { get; set; }
 
         public bool IsMtlsPopRequested { get; set; }
@@ -58,7 +52,6 @@ namespace Microsoft.Identity.Client.ApiConfig.Parameters
                      ForceRefresh: {ForceRefresh}
                      Resource: {Resource}
                      Claims: {!string.IsNullOrEmpty(Claims)}
-                     ClientClaims: {!string.IsNullOrEmpty(ClientClaims)}
                      RevokedTokenHash: {!string.IsNullOrEmpty(RevokedTokenHash)}
                      IsMtlsPopRequested: {IsMtlsPopRequested}
                      MtlsPopMinStrength: {MtlsPopMinStrength}
