+++ src/client/Microsoft.Identity.Client/ManagedIdentity/AbstractManagedIdentity.cs
@@ -15,8 +15,6 @@ using System.Text;
 using System.Security.Cryptography.X509Certificates;
 using System.Net.Security;
 using Microsoft.Identity.Client.Http.Retry;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
 
 namespace Microsoft.Identity.Client.ManagedIdentity
@@ -38,13 +36,6 @@ namespace Microsoft.Identity.Client.ManagedIdentity
             _sourceType = sourceType;
         }
 
-        // True only for the IMDSv1 source. IMDSv1 and IMDSv2 both report
-        // <see cref="ManagedIdentitySource.Imds"/> publicly, so this flag preserves the
-        // v1-specific MSIv1 claims validation without relying on the (folded) source label.
-        protected virtual bool RequiresMsiV1ClaimsValidation => false;
-
-        private const string XmsAzNwperimid = "xms_az_nwperimid";
-
         public virtual async Task<ManagedIdentityResponse> AuthenticateAsync(
             AcquireTokenForManagedIdentityParameters parameters,
             CancellationToken cancellationToken)
@@ -63,38 +54,6 @@ namespace Microsoft.Identity.Client.ManagedIdentity
             _isMtlsPopRequested = parameters.IsMtlsPopRequested;
 
             ManagedIdentityRequest request = await CreateRequestAsync(resource).ConfigureAwait(false);
-
-            // Forward client-originated claims to the correct location for IMDS/MSIv2 only.
-            // Other MI sources (App Service, Azure Arc, Service Fabric, etc.) do not have a
-            // confirmed contract for the "claims" parameter; fail fast rather than silently
-            // ignoring the value and polluting the cache with keys the endpoint never saw.
-            if (!string.IsNullOrEmpty(parameters.ClientClaims))
-            {
-                if (_sourceType != ManagedIdentitySource.Imds)
-                {
-                    throw new MsalClientException(
-                        MsalError.InvalidRequest,
-                        $"WithClaimsFromClient is only supported for IMDS-based managed identity sources. " +
-                        $"The detected source is {_sourceType}. " +
-                        "Only ManagedIdentitySource.Imds supports the 'claims' parameter.");
-                }
-
-                if (RequiresMsiV1ClaimsValidation)
-                {
-                    ValidateMsiv1Claims(parameters.ClientClaims);
-                }
-
-                if (request.Method == System.Net.Http.HttpMethod.Get)
-                {
-                    request.QueryParameters["claims"] = Uri.EscapeDataString(parameters.ClientClaims);
-                    _requestContext.Logger.Info("[Managed Identity] Adding client claims to IMDS request as query parameter.");
-                }
-                else
-                {
-                    request.BodyParameters["claims"] = parameters.ClientClaims;
-                    _requestContext.Logger.Info("[Managed Identity] Adding client claims to ESTS POST body.");
-                }
-            }
 
             // When IMDSv2 mints a binding certificate during this request (via CSR),
             // it's exposed via request.MtlsCertificate. Bubble it up so the request
@@ -370,27 +329,6 @@ namespace Microsoft.Identity.Client.ManagedIdentity
                 null);
 
             throw exception;
-        }
-
-        /// <summary>
-        /// MSIv1 (IMDS v1) only supports a single custom claim: <c>xms_az_nwperimid</c>.
-        /// Any other top-level key in the claims JSON will cause IMDS to return HTTP 400 Bad Request
-        /// with no useful diagnostic. Validate early so the caller gets a clear MSAL error.
-        /// </summary>
-        private static void ValidateMsiv1Claims(string claimsJson)
-        {
-            var parsed = ClaimsHelper.ParseClaimsOrThrow(claimsJson);
-            foreach (var kvp in parsed)
-            {
-                if (!string.Equals(kvp.Key, XmsAzNwperimid, StringComparison.Ordinal))
-                {
-                    throw new MsalClientException(
-                        MsalError.InvalidRequest,
-                        $"MSIv1 (IMDS v1) only supports the `{XmsAzNwperimid}` custom claim. " +
-                        $"The claims JSON contained the unsupported key `{kvp.Key}`. " +
-                        $"Remove all keys other than `{XmsAzNwperimid}` when using WithClaimsFromClient with MSIv1.");
-                }
-            }
         }
     }
 }
