+++ tests/Microsoft.Identity.Test.Unit/ManagedIdentityTests/ImdsV2Tests.cs
@@ -530,122 +530,6 @@ namespace Microsoft.Identity.Test.Unit.ManagedIdentityTests
             }
         }
 
-        // ---------------------------------------------------------------------------------
-        // Client-originated claims on the IMDSv2 mTLS-PoP path.
-        // Because the IMDSv2 token leg is delegated to MSAL's internal TokenClient exchange,
-        // client claims ride the shared ESTS-R POST body (merged with client capabilities),
-        // are keyed into the cache, and are NOT subject to the MSIv1 `xms_az_nwperimid` allowlist.
-        // ---------------------------------------------------------------------------------
-
-        // A full NSP-style claim. Unlike MSIv1, the IMDSv2 PoP path accepts arbitrary claim keys
-        // because the token leg is served by ESTS-R (same contract as confidential client).
-        private const string ClientClaims = @"{""xms_az_nwperimid"":{""essential"":true}}";
-        private const string OtherClientClaims = @"{""custom_claim"":{""values"":[""abc""]}}";
-
-        [TestMethod]
-        public async Task mTLSPop_WithClaimsFromClient_ForwardsClaimsInEstsBodyAsync()
-        {
-            using (new EnvVariableContext())
-            using (var httpManager = new MockHttpManager())
-            {
-                SetEnvironmentVariables(ManagedIdentitySource.Imds, TestConstants.ImdsEndpoint);
-
-                var managedIdentityApp = await CreateManagedIdentityAsync(
-                    httpManager,
-                    managedIdentityKeyType: ManagedIdentityKeyType.KeyGuard,
-                    withExperimentalFeatures: true).ConfigureAwait(false);
-
-                // The token-leg mock asserts the delegated ESTS-R POST body carries claims=<ClientClaims>.
-                // If the claims are not forwarded, the handler will not match and the test fails.
-                AddMocksToGetEntraToken(httpManager, expectedClaims: ClientClaims);
-
-                var result = await managedIdentityApp.AcquireTokenForManagedIdentity(ManagedIdentityTests.Resource)
-                    .WithMtlsProofOfPossession()
-                    .WithAttestationSupport()
-                    .WithClaimsFromClient(ClientClaims)
-                    .ExecuteAsync().ConfigureAwait(false);
-
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MTLSPoP, result.TokenType);
-                Assert.IsNotNull(result.BindingCertificate);
-                Assert.AreEqual(TokenSource.IdentityProvider, result.AuthenticationResultMetadata.TokenSource);
-            }
-        }
-
-        [TestMethod]
-        public async Task mTLSPop_WithClaimsFromClient_SameClaims_SecondCallFromCacheAsync()
-        {
-            using (new EnvVariableContext())
-            using (var httpManager = new MockHttpManager())
-            {
-                SetEnvironmentVariables(ManagedIdentitySource.Imds, TestConstants.ImdsEndpoint);
-
-                var managedIdentityApp = await CreateManagedIdentityAsync(
-                    httpManager,
-                    managedIdentityKeyType: ManagedIdentityKeyType.KeyGuard,
-                    withExperimentalFeatures: true).ConfigureAwait(false);
-
-                // Only one network mock — the second call with identical claims must be served from cache.
-                AddMocksToGetEntraToken(httpManager, expectedClaims: ClientClaims);
-
-                var result1 = await managedIdentityApp.AcquireTokenForManagedIdentity(ManagedIdentityTests.Resource)
-                    .WithMtlsProofOfPossession()
-                    .WithAttestationSupport()
-                    .WithClaimsFromClient(ClientClaims)
-                    .ExecuteAsync().ConfigureAwait(false);
-
-                var result2 = await managedIdentityApp.AcquireTokenForManagedIdentity(ManagedIdentityTests.Resource)
-                    .WithMtlsProofOfPossession()
-                    .WithAttestationSupport()
-                    .WithClaimsFromClient(ClientClaims)
-                    .ExecuteAsync().ConfigureAwait(false);
-
-                Assert.AreEqual(TokenSource.IdentityProvider, result1.AuthenticationResultMetadata.TokenSource,
-                    "First call should hit the network.");
-                Assert.AreEqual(TokenSource.Cache, result2.AuthenticationResultMetadata.TokenSource,
-                    "Second call with identical client claims must be served from cache.");
-            }
-        }
-
-        [TestMethod]
-        public async Task mTLSPop_WithClaimsFromClient_DifferentClaims_SeparateCacheEntriesAsync()
-        {
-            using (new EnvVariableContext())
-            using (var httpManager = new MockHttpManager())
-            {
-                SetEnvironmentVariables(ManagedIdentitySource.Imds, TestConstants.ImdsEndpoint);
-
-                var managedIdentityApp = await CreateManagedIdentityAsync(
-                    httpManager,
-                    managedIdentityKeyType: ManagedIdentityKeyType.KeyGuard,
-                    withExperimentalFeatures: true).ConfigureAwait(false);
-
-                // Two distinct network mocks — each claims value must produce a separate cache entry.
-                // The second acquire reuses the cached binding certificate, so it re-runs only the
-                // CSR-metadata + token leg (no /issuecredential), mirroring the cached-cert refresh path.
-                AddMocksToGetEntraToken(httpManager, expectedClaims: ClientClaims);
-                httpManager.AddMockHandler(MockHelpers.MockCsrResponse());
-                httpManager.AddMockHandler(MockHelpers.MockImdsV2EntraTokenRequestResponse(_identityLoggerAdapter, OtherClientClaims));
-
-                var result1 = await managedIdentityApp.AcquireTokenForManagedIdentity(ManagedIdentityTests.Resource)
-                    .WithMtlsProofOfPossession()
-                    .WithAttestationSupport()
-                    .WithClaimsFromClient(ClientClaims)
-                    .ExecuteAsync().ConfigureAwait(false);
-
-                var result2 = await managedIdentityApp.AcquireTokenForManagedIdentity(ManagedIdentityTests.Resource)
-                    .WithMtlsProofOfPossession()
-                    .WithAttestationSupport()
-                    .WithClaimsFromClient(OtherClientClaims)
-                    .ExecuteAsync().ConfigureAwait(false);
-
-                Assert.AreEqual(TokenSource.IdentityProvider, result1.AuthenticationResultMetadata.TokenSource,
-                    "First claims value should hit the network.");
-                Assert.AreEqual(TokenSource.IdentityProvider, result2.AuthenticationResultMetadata.TokenSource,
-                    "A different claims value must produce a separate cache entry and hit the network.");
-            }
-        }
-
         [TestMethod]
         public async Task mTLSPop_TokenLeg_InvalidClient_ReMintsBindingAndRetriesOnceAsync()
         {
